Fix/init bnd memory leak

+++ src/ast/free_ast.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free_ast.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: eduribei <eduribei@student.42.fr>          +#+  +:+       +#+        */
+/*   By: luide-ca <luide-ca@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/22 12:49:10 by luide-ca          #+#    #+#             */
-/*   Updated: 2025/05/27 21:55:39 by eduribei         ###   ########.fr       */
+/*   Updated: 2025/05/28 15:15:43 by luide-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 #include "../../include/execution.h"
 #include "../../include/parsing.h"
 
-void	ft_del_redir_node(void *content)
+void	ft_del_redirs_nodes(void *content)
 {
 	t_redirs_node	*redir_node;
 
@@ -24,9 +24,14 @@ void	ft_del_redir_node(void *content)
 		return ;
 	redir_node = (t_redirs_node *)content;
 	if (redir_node->type == HDC_R)
-		unlink(redir_node->name);
-	ft_free_and_null((void *)&redir_node->name);
-	//ft_free_and_null((void *)&redir_node);
+	{
+		if (redir_node->name != NULL)
+			unlink(redir_node->name);
+	}
+	if (redir_node->name)
+		ft_free_and_null((void *)&redir_node->name);
+	if (redir_node)
+		ft_free_and_null((void *)&redir_node);
 }
 
 void	free_block_node(void *ptr)
@@ -36,64 +41,16 @@ void	free_block_node(void *ptr)
 	blk = (t_block_node *)ptr;
 	if (!blk)
 		return ;
-	if (blk->output_lst)
-		ft_lstclear(&blk->output_lst, ft_del_redir_node);
-	if (blk->input_lst)
-		ft_lstclear(&blk->input_lst, ft_del_redir_node);
 	if (blk->redirs_lst)
-		ft_lstclear(&blk->redirs_lst, ft_del_redir_node);
+		ft_lstclear(&blk->redirs_lst, ft_del_redirs_nodes);
+	if (blk->output_lst)
+		ft_lstclear(&blk->output_lst, NULL);
+	if (blk->input_lst)
+		ft_lstclear(&blk->input_lst, NULL);
 	if (blk->cmd_arr)
 		ft_free_and_null_str_array(&blk->cmd_arr);
 	free(blk);
 }
-
-//Free AST memory
-// void	ft_free_ast(t_ast_node *node)
-// {
-// 	t_block_node	*blk;
-// 	t_pipe_info		*pipe;
-// 	t_logical_data	*log;
-// 	t_subshell_data	*sub;
-
-// 	if (!node)
-// 		return ;
-// 	if (node->type == NODE_COMMAND)
-// 	{
-// 		blk = node->block_node;
-// 		if (blk)
-// 			free_block_node(blk);
-// 	}
-// 	else if (node->type == NODE_PIPELINE)
-// 	{
-// 		pipe = node->pipeline;
-// 		if (pipe)
-// 		{
-// 			ft_lstiter(pipe->cmds, (void (*)(void *))ft_free_ast);
-// 			ft_lstclear(&pipe->cmds, NULL);
-// 			free(pipe);
-// 		}
-// 	}
-// 	else if (node->type == NODE_LOGICAL)
-// 	{
-// 		log = node->logical;
-// 		if (log)
-// 		{
-// 			ft_free_ast(log->left);
-// 			ft_free_ast(log->right);
-// 			free(log);
-// 		}
-// 	}
-// 	else if (node->type == NODE_SUBSHELL)
-// 	{
-// 		sub = node->subshell;
-// 		if (sub)
-// 		{
-// 			ft_free_ast(sub->body);
-// 			free(sub);
-// 		}
-// 	}
-// 	free(node);
-// }
 
 void	ft_free_ast(t_ast_node **node)
 {
@@ -115,8 +72,7 @@ void	ft_free_ast(t_ast_node **node)
 		pipe = (*node)->pipeline;
 		if (pipe)
 		{
-			ft_lstiter(pipe->cmds, (void (*)(void *))ft_free_ast);
-			ft_lstclear(&pipe->cmds, NULL);
+			ft_lstclear(&pipe->cmds, free_block_node);
 			free(pipe);
 		}
 	}
